split scopes into descriptions/validation and action
