+++ pkg/oauth/apiserver/registry/oauthauthorizetoken/strategy.go
@@ -3,13 +3,14 @@ package oauthauthorizetoken
 import (
 	"context"
 
+	"github.com/openshift/origin/pkg/authorization/authorizer/scopelibrary"
+
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/util/validation/field"
 	"k8s.io/apiserver/pkg/registry/rest"
 	"k8s.io/kubernetes/pkg/api/legacyscheme"
 
-	scopeauthorizer "github.com/openshift/origin/pkg/authorization/authorizer/scope"
 	oauthapi "github.com/openshift/origin/pkg/oauth/apis/oauth"
 	"github.com/openshift/origin/pkg/oauth/apis/oauth/validation"
 	"github.com/openshift/origin/pkg/oauth/apiserver/registry/oauthclient"
@@ -61,7 +62,7 @@ func (s strategy) Validate(ctx context.Context, obj runtime.Object) field.ErrorL
 	if err != nil {
 		return append(validationErrors, field.InternalError(field.NewPath("clientName"), err))
 	}
-	if err := scopeauthorizer.ValidateScopeRestrictions(client, token.Scopes...); err != nil {
+	if err := scopelibrary.ValidateScopeRestrictions(client, token.Scopes...); err != nil {
 		return append(validationErrors, field.InternalError(field.NewPath("clientName"), err))
 	}
